storage: move Y2Storage conversions
