+++ service/test/agama/storage/volume_test.rb
@@ -22,6 +22,7 @@
 require_relative "../../test_helper"
 require "agama/config"
 require "agama/storage/volume"
+require "y2storage/volume_specification"
 
 describe Agama::Storage::Volume do
   describe ".new_from_json" do
@@ -48,6 +49,15 @@ describe Agama::Storage::Volume do
     it "generates a JSON hash according to schema" do
       result = volume.to_json_settings
       expect(result).to be_a(Hash)
+    end
+  end
+
+  describe "#to_y2storage" do
+    let(:volume) { Agama::Storage::Volume.new("/test") }
+
+    it "generates a Y2Storage volume spec" do
+      result = volume.to_y2storage
+      expect(result).to be_a(Y2Storage::VolumeSpecification)
     end
   end
 end
